additional HealthKit data

+++ AccessibleScale/Helpers/GATTBodyCompositionMeasurement.swift
@@ -13,13 +13,13 @@ struct GATTBodyCompositionMeasurement {
     let fatPercentage: Double
     let timestamp: Date?
     let userID: UInt8?
-    let basalMetabolism: UInt16?
+    let basalMetabolism: Double?
     let musclePercentage: Double?
     let muscleMass: Double?
     let fatFreeMass: Double?
     let softLeanMass: Double?
     let bodyWaterMass: Double?
-    let impedance: UInt16?
+    let impedance: Double?
     let weight: Double?
     let height: Double?
     let multiplePacket: Bool
@@ -53,13 +53,13 @@ struct GATTBodyCompositionMeasurement {
 
         timestamp =        (GATTUtil.flag(flag, timestampBit) && index + 7 < length) ?        GATTDate(data: data, start: &index).date : nil
         userID =           (GATTUtil.flag(flag, userIDBit) && index + 1 < length) ?           GATTUtil.uint8(data, &index) : nil
-        basalMetabolism =  (GATTUtil.flag(flag, basalMetabolismBit) && index + 2 < length) ?  GATTUtil.uint16(data, &index) : nil
+        basalMetabolism =  (GATTUtil.flag(flag, basalMetabolismBit) && index + 2 < length) ?  Double(GATTUtil.uint16(data, &index)) : nil
         musclePercentage = (GATTUtil.flag(flag, musclePercentageBit) && index + 2 < length) ? Double(GATTUtil.uint16(data, &index)) * percentFactor : nil
         muscleMass =       (GATTUtil.flag(flag, muscleMassBit) && index + 2 < length) ?       Double(GATTUtil.uint16(data, &index)) * massFactor : nil
         fatFreeMass =      (GATTUtil.flag(flag, fatFreeMassBit) && index + 2 < length) ?      Double(GATTUtil.uint16(data, &index)) * massFactor : nil
         softLeanMass =     (GATTUtil.flag(flag, softLeanMassBit) && index + 2 < length) ?     Double(GATTUtil.uint16(data, &index)) * massFactor : nil
         bodyWaterMass =    (GATTUtil.flag(flag, bodyWaterMassBit) && index + 2 < length) ?    Double(GATTUtil.uint16(data, &index)) * massFactor : nil
-        impedance =        (GATTUtil.flag(flag, impedanceBit) && index + 2 < length) ?        GATTUtil.uint16(data, &index) : nil
+        impedance =        (GATTUtil.flag(flag, impedanceBit) && index + 2 < length) ?        Double(GATTUtil.uint16(data, &index)) : nil
         weight =           (GATTUtil.flag(flag, weightBit) && index + 2 < length) ?           Double(GATTUtil.uint16(data, &index)) * massFactor : nil
         height =           (GATTUtil.flag(flag, heightBit) && index + 2 < length) ?           Double(GATTUtil.uint16(data, &index)) * heightFactor : nil
     }
